+++ dsr_agents/include/dsr_agents/docking_agent.hpp
@@ -16,57 +16,37 @@
 #ifndef DSR_AGENTS__DOCKING_AGENT_HPP_
 #define DSR_AGENTS__DOCKING_AGENT_HPP_
 
-// C++
-#include <memory>
-#include <string>
-#include <vector>
-
 // ROS
-#include "rclcpp/rclcpp.hpp"
-#include "rclcpp_action/rclcpp_action.hpp"
 #include "opennav_docking_msgs/action/dock_robot.hpp"
-#include "opennav_docking_msgs/action/undock_robot.hpp"
 
 // DSR
-#include "dsr/api/dsr_api.h"
-#include "dsr_util/agent_node.hpp"
+#include "dsr_util/action_agent.hpp"
 
 namespace dsr_agents
 {
 
-class DockingAgent : public dsr_util::AgentNode
+/**
+ * @class dsr_agents::DockingAgent
+ * @brief Agent to dock the robot in a charging station.
+ */
+class DockingAgent : public dsr_util::ActionAgent<opennav_docking_msgs::action::DockRobot>
 {
 public:
+  /**
+   * @brief Construct a new Docking Agent object.
+   *
+   * @param options The options for the ROS node.
+   */
   explicit DockingAgent(const rclcpp::NodeOptions & options = rclcpp::NodeOptions());
 
-private:
-  using Dock = opennav_docking_msgs::action::DockRobot;
-  using Undock = opennav_docking_msgs::action::UndockRobot;
-  using GoalHandleDock = rclcpp_action::ClientGoalHandle<Dock>;
-  using GoalHandleUndock = rclcpp_action::ClientGoalHandle<Undock>;
+protected:
 
-  rclcpp_action::Client<Dock>::SharedPtr dock_client_;
-  rclcpp_action::Client<Undock>::SharedPtr undock_client_;
-  std::shared_ptr<GoalHandleDock> goal_handle_;
-
-  void cancel_action();
-  void start_docking();
-  void start_undocking();
-
-  // Docking action callbacks
-  void dock_goal_response_callback(const GoalHandleDock::SharedPtr & goal_handle);
-  void dock_feedback_callback(
-    GoalHandleDock::SharedPtr, const std::shared_ptr<const Dock::Feedback> feedback);
-  void dock_result_callback(const GoalHandleDock::WrappedResult & result);
-
-  // Undocking action callbacks
-  void undock_goal_response_callback(const GoalHandleUndock::SharedPtr & goal_handle);
-  void undock_feedback_callback(
-    GoalHandleUndock::SharedPtr, const std::shared_ptr<const Undock::Feedback> feedback);
-  void undock_result_callback(const GoalHandleUndock::WrappedResult & result);
-
-  // DSR callbacks
-  void edge_updated(std::uint64_t from, std::uint64_t to, const std::string & type) override;
+  /**
+   * @brief Get the dock_id from the DSR node.
+   * 
+   * @param action_node The DSR node with the dock_id information.
+   */
+  void get_goal_from_dsr(DSR::Node action_node) override;
 };
 
 }  // namespace dsr_agents