+++ dsr_agents/CHANGELOG.rst
@@ -28,6 +28,7 @@ X.X.X (XX-XX-2024)
 * Converted agents to composables.
 * Added unit tests.
 * Added on_cleanup function to delete nodes.
+* Navigation and docking agents now inherit from action_node.
 
 0.7.0 (19-06-2024)
 ------------------